# server/api.yaml
openapi: "3.0.3"
info:
  title: Call Center Broker API
  version: "1.0.0"
  description: >
    Broker API that dispatches client leads to agent mobile devices.
    Agents authenticate, receive clients to call, and submit call outcomes.
  license:
    name: Proprietary
servers:
  - url: https://api.server.test/api/v1
  - url: http://localhost:3000/api/v1
    description: Local dev v1
  - url: https://api.server.test/api/v2
  - url: http://localhost:3000/api/v2
    description: Local dev v2 (clients; v1 keeps /owners path for compat)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CallStatus:
      type: string
      enum: [answered, no_answer, busy, failed, callback, not_interested]
    ClientType:
      type: string
      enum: [OWNER, LEAD, BOTH, UNKNOWN]
      description: UNKNOWN is the API representation of a NULL type in the database
    ClientPhone:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          example: "+201001234567"
    ClientInfo:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
          example: "national_id"
        value:
          type: string
          example: "29901011234567"
    Client:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: "Mohamed Ali"
        type:
          $ref: "#/components/schemas/ClientType"
        next_dial_at:
          type: string
          format: date-time
          nullable: true
        agent_id:
          type: integer
          nullable: true
        phones:
          type: array
          items:
            $ref: "#/components/schemas/ClientPhone"
        info:
          type: array
          items:
            $ref: "#/components/schemas/ClientInfo"
...
    Project:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
          example: "Q3 Collections Campaign"
        description:
          type: string
          nullable: true
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          example: "eyJhbGci..."
        user:
          $ref: "#/components/schemas/User"
...
    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          type: string
        has_profile_image:
          type: boolean
          description: Whether the user has a profile image stored
        is_online:
          type: boolean
    ProjectRef:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    CallDetailRecord:
      type: object
      required: [id, client_id, agent_id, status, time]
      properties:
        id:
          type: integer
          format: int64
        client_id:
          type: integer
          format: int64
        agent_id:
          type: integer
        status:
          $ref: "#/components/schemas/CallStatus"
        time:
          type: string
          format: date-time
        duration:
          type: integer
          description: Call duration in seconds
          nullable: true
        agent_notes:
          type: string
          nullable: true
        projects:
          type: array
          items:
            $ref: "#/components/schemas/ProjectRef"
    NextClientResponse:
      type: object
      required: [client, calls]
      properties:
        client:
          $ref: "#/components/schemas/Client"
        owner:
          $ref: "#/components/schemas/Client"
          description: Deprecated v1 alias of client; both are returned for compatibility
        calls:
          type: array
          items:
            $ref: "#/components/schemas/CallDetailRecord"
...
    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
          example: "completed"
        count:
          type: integer
          example: 42
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "OWNER_NOT_FOUND"
        message:
          type: string
          example: "No client with that ID exists"
        details:
          type: object
          nullable: true
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request body or params failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: The request references a resource that does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request conflicts with the current state of the resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
paths:
  /login:
    post:
      summary: Agent / admin login
      operationId: login
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /register:
    post:
      summary: Register a new user (auto-login)
      operationId: register
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
                name:
                  type: string
                phone:
                  type: string
                role:
                  type: string
      responses:
        "201":
          description: User created and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /logout:
    post:
      summary: Invalidate the current session
      operationId: logout
      tags: [auth]
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me:
    get:
      summary: Get the currently authenticated user
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /projects:
    get:
      summary: List projects visible to the current user
      operationId: listProjects
      tags: [projects]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Create a new project
      operationId: createProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /projects/{projectId}:
    get:
      summary: Get a single project
      operationId: getProject
      tags: [projects]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update a project
      operationId: updateProject
      tags: [projects]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete a project (admin only)
      operationId: deleteProject
      tags: [projects]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Project deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /calls:
    get:
      summary: List call history
      operationId: listCalls
      tags: [calls]
      parameters:
        - name: client_id
          in: query
          schema:
            type: integer
            format: int64
        - name: agent_id
          in: query
          schema:
            type: integer
        - name: project_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CallStatus"
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated call records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CallDetailRecord"
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Submit a call outcome
      operationId: submitCall
      tags: [calls]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, status, time]
              properties:
                client_id:
                  type: integer
                  format: int64
                project_id:
                  type: integer
                  nullable: true
                status:
                  $ref: "#/components/schemas/CallStatus"
                time:
                  type: string
                  format: date-time
                duration:
                  type: integer
                  description: Duration in seconds
                agent_notes:
                  type: string
                  maxLength: 2000
                next_dial_at:
                  type: string
                  format: date-time
                  description: Callback time (only when status is callback)
      responses:
        "201":
          description: Call record saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallDetailRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
  /calls/next:
    get:
      summary: Dispatch the next client to call for this agent
      operationId: getNextClient
      tags: [calls]
      parameters:
        - name: project_id
          in: query
          required: false
          schema:
            type: integer
        - name: assigned_only
          in: query
          description: If true, only return clients assigned to the calling agent.
          schema:
            type: boolean
        - name: type
          in: query
          description: Filter by client type. If omitted, returns all types.
          schema:
            $ref: "#/components/schemas/ClientType"
      responses:
        "200":
          description: Next client to call with past call history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NextClientResponse"
        "204":
          description: No clients available right now
        "401":
          $ref: "#/components/responses/Unauthorized"
  /calls/calling:
    post:
      summary: Notify the server that the agent is currently calling a client
      operationId: notifyCalling
      tags: [calls]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id]
              properties:
                client_id:
                  type: integer
                  format: int64
                project_id:
                  type: integer
                  nullable: true
                client_number:
                  type: string
                  description: The specific number being called (optional)
      responses:
        "200":
          description: Notification received
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /calls/{callId}:
    get:
      summary: Get a single call record
      operationId: getCall
      tags: [calls]
      parameters:
        - name: callId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Call detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallDetailRecord"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a call record (admin only)
      operationId: deleteCall
      tags: [calls]
      parameters:
        - name: callId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Call record deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /calls/statuses:
    get:
      summary: Get call status counts with optional time filtering
      operationId: getCallStatuses
      tags: [calls]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Start of time window (ISO 8601)
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: End of time window (ISO 8601)
      responses:
        "200":
          description: List of statuses with counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"
        "401":
          $ref: "#/components/responses/Unauthorized"
...